pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sewa"
version = "0.1.0"
description = "Book Request Automation System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "phonenumbers>=8.13.0",
    "googlemaps>=4.10.0",
    "twilio>=8.10.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.17.0",
    "streamlit-option-menu>=0.3.6"
]
